pyproject.toml
[project]
name = "kanby"
version = "0.3.0"
description = "Conversational Kanban co-pilot"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICEN[CS]E*"]
authors = [
    { name = "Timur Çakmakoğlu", email = "timurcakmakoglu@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.18.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
api = [
    "psycopg[binary]>=3.2.13",
]
dev = [
    "google-adk[eval]>=1.18.0",
    "pre-commit>=4.4.0",
    "ruff>=0.14.5",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812"]

[tool.ruff.lint.pydocstyle]
convention = "google"
